auto_ria_scraper/selenium/requirements.txt
selenium
psycopg2-binary
webdriver-manager